Исправил проблему, что второй браузер берет данные из БД, а не из redis для GET /facilities
